move date range format to separate file
